ADDED: Asynchronous parsing of SVGKImage's ADDED: approximate progress-bar for loading large SVGKImage's (estimates by looking at how many bytes have been read from the file and XML parsed)

+++ Source/Parsers/SVGKParser.m
@@ -243,7 +243,7 @@ readPacket(char *mem, int size) {
 			{
 				if( source.approximateLengthInBytesOr0 > 0 )
 				{
-					currentParseRun.parseProgressFractionApproximate = totalBytesRead / source.approximateLengthInBytesOr0;
+					currentParseRun.parseProgressFractionApproximate = totalBytesRead / (double) source.approximateLengthInBytesOr0;
 				}
 				else
 					currentParseRun.parseProgressFractionApproximate = 0;
